server/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.1
SQLAlchemy==1.4.25
Werkzeug==2.0.2
